money: use exact DECIMAL for price and net_notional, not floating point
Prices are money and must not round-trip through IEEE754. Postgres already stores
price as NUMERIC(12,4); the pipeline was downgrading it to DOUBLE in every engine,
and net_notional = SUM(quantity * price) accumulates that error across millions of
rows, so the gold layer would drift in a way no test would obviously catch.

- Debezium decimal.handling.mode: string (exact decimal on the wire). `precise`
  emits base64 Connect Decimal bytes the JSON consumers cannot decode; `double`
  parses but is lossy, it was a stop-gap I set earlier and is now reverted.
- price -> DECIMAL(12,4) everywhere, matching the source column exactly: Delta and
  Iceberg DDL, Flink Fluss/Paimon DDL and the debezium-json ROW types.
- net_notional -> DECIMAL(38,4). Wider than price because it is a running sum of
  quantity x price over millions of rows and needs the integer headroom.
- Spark bronze jobs parse price as StringType and cast to DECIMAL(12,4), so the
  exact value from Debezium is never routed through a float on the way in.

Applies to every engine equally, so the comparison stays like-for-like.

NOTE: this changes table schemas, so existing tables must be recreated before the
next run, the change takes effect on redeploy, not on the currently running one.

+++ jobs/_shared/iceberg_tables.py
@@ -26,12 +26,12 @@ _STATEMENTS = [
     f"CREATE NAMESPACE IF NOT EXISTS {_CAT}.gold",
     f"""CREATE TABLE IF NOT EXISTS {_CAT}.bronze.trades_spark (
         op STRING, trade_id BIGINT, account_id BIGINT, symbol STRING, side STRING,
-        quantity INT, price DOUBLE, executed_at TIMESTAMP, event_ts TIMESTAMP,
+        quantity INT, price DECIMAL(12,4), executed_at TIMESTAMP, event_ts TIMESTAMP,
         ingest_ts TIMESTAMP, kafka_offset BIGINT, kafka_partition INT)
       USING iceberg PARTITIONED BY (days(executed_at)) TBLPROPERTIES ({_APPEND_PROPS})""",
     f"""CREATE TABLE IF NOT EXISTS {_CAT}.silver.trades_spark (
         trade_id BIGINT, account_id BIGINT, symbol STRING, side STRING, quantity INT,
-        price DOUBLE, executed_at TIMESTAMP, event_ts TIMESTAMP, ingest_ts TIMESTAMP)
+        price DECIMAL(12,4), executed_at TIMESTAMP, event_ts TIMESTAMP, ingest_ts TIMESTAMP)
       USING iceberg PARTITIONED BY (days(executed_at)) TBLPROPERTIES ({_APPEND_PROPS})""",
     f"""CREATE TABLE IF NOT EXISTS {_CAT}.silver.accounts_spark (
         account_id BIGINT NOT NULL, name STRING, country STRING, tier STRING,
@@ -39,7 +39,7 @@ _STATEMENTS = [
       USING iceberg TBLPROPERTIES ({_MOR_PROPS})""",
     f"""CREATE TABLE IF NOT EXISTS {_CAT}.gold.open_positions_spark (
         account_id BIGINT NOT NULL, symbol STRING NOT NULL, net_quantity BIGINT,
-        net_notional DOUBLE, trade_count BIGINT, status STRING, country STRING,
+        net_notional DECIMAL(38,4), trade_count BIGINT, status STRING, country STRING,
         tier STRING, commit_ts TIMESTAMP)
       USING iceberg PARTITIONED BY (symbol) TBLPROPERTIES ({_MOR_PROPS})""",
 ]
